+++ javascript/modules/client/test/integration/client-addressList.test.ts
@@ -18,6 +18,7 @@ import { GraphQLClient } from "graphql-request";
 import {
   ExecuteProposalStep,
   IAddressListPluginInstall,
+  ICanVoteParams,
   ICreateProposalParams,
   IExecuteProposalParams,
   IPluginSettings,
@@ -358,6 +359,21 @@ describe("Client", () => {
             );
         }
       }
+    });
+  });
+
+  describe("Can vote", () => {
+    it("Should check if an user can vote in an Address List proposal", async () => {
+      const context = new ContextPlugin(contextParamsLocalChain);
+      const client = new ClientAddressList(context);
+
+      const params: ICanVoteParams = {
+        address: "0x1234567890123456789012345678901234567890",
+        proposalId: "0x1234567890123456789012345678901234567890_0x1",
+        pluginAddress: "0x1234567890123456789012345678901234567890",
+      };
+      const canVote = await client.methods.canVote(params);
+      expect(typeof canVote).toBe("boolean");
     });
   });