+++ javascript/modules/client/CHANGELOG.md
@@ -17,7 +17,8 @@ TEMPLATE:
 -->
 
 ## [UPCOMING]
-
+### Added
+- Adding the `canVote` getter to the address list and ERC20 clients
 ## 0.7.0-alpha
 On 2022-09-21 14:50:29
 ### Added