+++ javascript/modules/common/CHANGELOG.md
@@ -17,7 +17,8 @@ TEMPLATE:
 -->
 
 ## [UPCOMING]
-
+### Added
+- New error types
 ## 0.4.0-alpha
 On 2022-09-21 14:49:57
 ### Added